requirements.txt
requests
mysql-connector
spacy
plotly
psutil
